feature: Added `PostBox` as part of a link abstraction on top of (#1)
`Worker`. This moves the `work` call to be asynchronous in nature - the
remote worker will send a reply first confirming the job had been
received; upon completion of task, the remote worker will then open a
different connection back to the host to deliver the work result.

This allows all result waiting to be unified in one listening port, and
permits further link abstractions such as message deduplication etc,
which `PostBox` already supports.